requirements.txt
mxnet==1.5.0
gluoncv==0.4.0
numpy
Pillow